SDK search directory moved
