+++ frontend/packages/multicluster-sdk/src/api/index.ts
@@ -4,7 +4,7 @@ export * from './FleetResourceLink'
 export * from './useFleetK8sAPIPath'
 export * from './useFleetK8sWatchResource'
 export * from './useFleetClusterNames'
-export * from './search/useFleetSearchPoll'
+export * from './useFleetSearchPoll'
 export * from './apiRequests'
 export * from './useHubClusterName'
 export * from './useFleetPrometheusPoll'
