+++ frontend/packages/multicluster-sdk/src/internal/search/search-sdk.ts
