+++ frontend/packages/multicluster-sdk/src/internal/search/searchQueries.graphql
