+++ frontend/packages/multicluster-sdk/src/internal/search/searchUtils.ts
